refactor(daemon): move DaemonLaunchDescriptor to daemon-protocol
`daemon-launch.json` is written by the gradle plugin and read by the daemon
JVM, the CLI doctor, the VS Code extension and serve. A format four programs
agree on is the definition of a protocol; the descriptor sat in `:daemon:core`
only because that is where the daemon was written. It is 95 lines of
`@Serializable` data with no dependency on anything in that module.

This is the last of serve's `daemon-core` imports that was a shape rather
than behaviour. What remains is four, in two groups that want different
answers: `DeviceDimensions` / `frameDpOverriddenBy` (a device catalog and
arithmetic over it — a contract by the same reasoning as
`data-pseudolocale-core`), and `BtaCompileSession` / `DiagnosticCollector`
(the playground's in-process compile, which is genuine coupling and a product
question rather than a module-layout one).

The package moves with it, so this is not free the way #4670 was: sixteen
files carried the fully-qualified name. Rewritten by exact FQN rather than by
symbol name, because a name-based sweep on this codebase has produced
plausible, compiling, wrong imports before. No file declares
`ee.schimke.composeai.daemon` and uses the type unqualified, so nothing is
hiding behind same-package resolution — checked directly rather than inferred
from a green compile.

Verified: nine modules compile, four test suites green, the contract probe
resolves 17 contracts with 0 leaks, both seam gates OK. `:daemon:android` has
no reference to the type at all, which is as well since this sandbox has no
Android SDK to compile it with.
