+++ cli/serve/src/test/kotlin/ee/schimke/composeai/cli/serve/ServeBundleDaemonWeightTest.kt
@@ -1,6 +1,6 @@
 package ee.schimke.composeai.cli.serve
 
-import ee.schimke.composeai.daemon.DaemonLaunchDescriptor
+import ee.schimke.composeai.daemon.protocol.DaemonLaunchDescriptor
 import java.io.File
 import kotlin.test.Test
 import kotlin.test.assertEquals
