+++ daemon/core/api/core.api
@@ -139,62 +139,6 @@ public final class ee/schimke/composeai/daemon/CompositeDataProductRegistry : ee
 	public fun renderModeFor (Ljava/lang/String;)Ljava/lang/String;
 }
 
-public final class ee/schimke/composeai/daemon/DaemonLaunchDescriptor {
-	public static final field Companion Lee/schimke/composeai/daemon/DaemonLaunchDescriptor$Companion;
-	public static final field SANDBOX_COUNT_PROP Ljava/lang/String;
-	public fun <init> (ILjava/lang/String;Ljava/lang/String;ZLjava/lang/String;Ljava/lang/String;Ljava/util/List;Ljava/util/List;Ljava/util/Map;Ljava/lang/String;Ljava/lang/String;Ljava/util/List;Ljava/lang/Long;)V
-	public synthetic fun <init> (ILjava/lang/String;Ljava/lang/String;ZLjava/lang/String;Ljava/lang/String;Ljava/util/List;Ljava/util/List;Ljava/util/Map;Ljava/lang/String;Ljava/lang/String;Ljava/util/List;Ljava/lang/Long;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public final fun component1 ()I
-	public final fun component10 ()Ljava/lang/String;
-	public final fun component11 ()Ljava/lang/String;
-	public final fun component12 ()Ljava/util/List;
-	public final fun component13 ()Ljava/lang/Long;
-	public final fun component2 ()Ljava/lang/String;
-	public final fun component3 ()Ljava/lang/String;
-	public final fun component4 ()Z
-	public final fun component5 ()Ljava/lang/String;
-	public final fun component6 ()Ljava/lang/String;
-	public final fun component7 ()Ljava/util/List;
-	public final fun component8 ()Ljava/util/List;
-	public final fun component9 ()Ljava/util/Map;
-	public final fun copy (ILjava/lang/String;Ljava/lang/String;ZLjava/lang/String;Ljava/lang/String;Ljava/util/List;Ljava/util/List;Ljava/util/Map;Ljava/lang/String;Ljava/lang/String;Ljava/util/List;Ljava/lang/Long;)Lee/schimke/composeai/daemon/DaemonLaunchDescriptor;
-	public static synthetic fun copy$default (Lee/schimke/composeai/daemon/DaemonLaunchDescriptor;ILjava/lang/String;Ljava/lang/String;ZLjava/lang/String;Ljava/lang/String;Ljava/util/List;Ljava/util/List;Ljava/util/Map;Ljava/lang/String;Ljava/lang/String;Ljava/util/List;Ljava/lang/Long;ILjava/lang/Object;)Lee/schimke/composeai/daemon/DaemonLaunchDescriptor;
-	public fun equals (Ljava/lang/Object;)Z
-	public final fun getClasspath ()Ljava/util/List;
-	public final fun getEnabled ()Z
-	public final fun getHardTtlSeconds ()Ljava/lang/Long;
-	public final fun getJailCommand ()Ljava/util/List;
-	public final fun getJavaLauncher ()Ljava/lang/String;
-	public final fun getJvmArgs ()Ljava/util/List;
-	public final fun getMainClass ()Ljava/lang/String;
-	public final fun getManifestPath ()Ljava/lang/String;
-	public final fun getModulePath ()Ljava/lang/String;
-	public final fun getSchemaVersion ()I
-	public final fun getSystemProperties ()Ljava/util/Map;
-	public final fun getVariant ()Ljava/lang/String;
-	public final fun getWorkingDirectory ()Ljava/lang/String;
-	public fun hashCode ()I
-	public final fun jailed (Ljava/util/List;Ljava/lang/Long;)Lee/schimke/composeai/daemon/DaemonLaunchDescriptor;
-	public fun toString ()Ljava/lang/String;
-	public final fun withSandboxCount (I)Lee/schimke/composeai/daemon/DaemonLaunchDescriptor;
-}
-
-public final synthetic class ee/schimke/composeai/daemon/DaemonLaunchDescriptor$$serializer : kotlinx/serialization/internal/GeneratedSerializer {
-	public static final field INSTANCE Lee/schimke/composeai/daemon/DaemonLaunchDescriptor$$serializer;
-	public final fun childSerializers ()[Lkotlinx/serialization/KSerializer;
-	public final fun deserialize (Lkotlinx/serialization/encoding/Decoder;)Lee/schimke/composeai/daemon/DaemonLaunchDescriptor;
-	public synthetic fun deserialize (Lkotlinx/serialization/encoding/Decoder;)Ljava/lang/Object;
-	public final fun getDescriptor ()Lkotlinx/serialization/descriptors/SerialDescriptor;
-	public final fun serialize (Lkotlinx/serialization/encoding/Encoder;Lee/schimke/composeai/daemon/DaemonLaunchDescriptor;)V
-	public synthetic fun serialize (Lkotlinx/serialization/encoding/Encoder;Ljava/lang/Object;)V
-	public fun typeParametersSerializers ()[Lkotlinx/serialization/KSerializer;
-}
-
-public final class ee/schimke/composeai/daemon/DaemonLaunchDescriptor$Companion {
-	public final fun parse (Ljava/lang/String;)Lee/schimke/composeai/daemon/DaemonLaunchDescriptor;
-	public final fun serializer ()Lkotlinx/serialization/KSerializer;
-}
-
 public final class ee/schimke/composeai/daemon/DaemonVersion {
 	public static final field INSTANCE Lee/schimke/composeai/daemon/DaemonVersion;
 	public final fun getValue ()Ljava/lang/String;
